requirements.txt
pandas==2.2.3
requests==2.32.3
python-dotenv==1.0.1
PyYAML==6.0.2
langchain==0.3.1
langchain-community==0.3.1
langchain-google-community==2.0.2
langchain-core==0.3.6
openai==1.51.0
beautifulsoup4==4.12.3
Flask==3.0.3
Flask-Cors==5.0.0
scikit-learn==1.5.2
tensorflow==2.18.0
matplotlib==3.9.2